requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-jose
python-multipart
psycopg2-binary
python-dotenv
email-validator
alembic
bcrypt==4.0.1
passlib[bcrypt]





